fix: add prsima and move supabase
