+++ src/hooks/useSupabase.ts
@@ -1,4 +1,4 @@
-import { supabase } from "@/lib/supabase";
+import { supabase } from "@/services/supabase";
 
 export const useSupabase = () => {
   return { supabase };
